Fix:#64 - fix error of msw
